ai_personal_assistant backend/requirements.txt
# FastAPI Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# Google Gemini AI - Latest Version
google-generativeai>=0.8.0

# Google APIs for Calendar, Meet
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Voice Processing - TTS with Edge-TTS and gTTS fallback
SpeechRecognition==3.10.1
edge-tts==6.1.9
gTTS==2.5.1
pydub==0.25.1

# Search APIs
duckduckgo-search==4.4.3
httpx==0.26.0

# Async HTTP
aiohttp==3.9.3
aiofiles==23.2.1

# Email
aiosmtplib==3.0.1

# Scheduler for reminders
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Audio Processing Dependencies
# Note: FFmpeg must be installed on the system for pydub
# Windows: choco install ffmpeg
# Linux: apt-get install ffmpeg
# macOS: brew install ffmpeg
